backend/pyproject.toml
[project]
name = "threat-analyst"
version = "0.1.0"
description = "Multi-agent cybersecurity threat analyst with MITRE ATT&CK RAG"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.4",
    "langchain>=0.3",
    "langchain-core>=0.3",
    "langchain-community>=0.3",
    "langchain-ollama>=0.3",
    "qdrant-client>=1.12",
    "FlagEmbedding>=1.3",
    "sentence-transformers>=3.0",
    "transformers>=4.45,<5.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "sse-starlette>=2.0",
    "mitreattack-python>=1.0",
    "stix2>=3.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "aiohttp>=3.10",
    "slowapi>=0.1.9",
    "cachetools>=5.0",
    "aiosqlite>=0.20",
    "prometheus-fastapi-instrumentator>=7.0",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "aiosmtplib>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-httpx>=0.35",
    "ruff>=0.8",
]
groq = [
    "langchain-groq>=0.2",
]
finetune = [
    "transformers>=4.45",
    "peft>=0.13",
    "trl>=0.12",
    "bitsandbytes>=0.44",
    "accelerate>=1.0",
    "datasets>=3.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["agents*", "api*", "ingestion*", "rag*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
